大致完成pms

+++ mc-modules/mc-wms/mc-wms-server/src/main/java/com/qcmoke/wms/entity/ProductStock.java
@@ -1,17 +1,18 @@
 package com.qcmoke.wms.entity;
 
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
+import java.io.Serializable;
+import java.util.Date;
+
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author qcmoke
@@ -23,7 +24,7 @@ import lombok.experimental.Accessors;
 @TableName("t_product_stock")
 public class ProductStock implements Serializable {
 
-private static final long serialVersionUID=1L;
+    private static final long serialVersionUID = 1L;
 
     /**
      * 产品库存编号
@@ -49,7 +50,7 @@ private static final long serialVersionUID=1L;
     /**
      * 库存量
      */
-    private Long count;
+    private Double count;
 
     /**
      * 创建时间
